# [REG-1709] Remove behaviour state duplication + improve performance