… src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/ScreenRecorder.cs