- Removes duplication of behaviour/collider/rigidbody states on parent objects vs child objects
  - Hand compared old and new states to ensure that things are showing up only once and that everything is still showing up for BossRoom
  - This change improved the performance of `ToJsonString` which runs only on KeyFrame ticks; the improvement is due to the reduction in behaviour json being written to the state (behaviours are the most expensive thing to convert to json)
    - new - 11.78ms , old - 16.31ms 
  - This change also reduced the GC allocs for a KeyFrame tick due to the reduction in state being recorded 
- Adapts prior state code paths to use a dictionary instead of a list to further improve performance with high object counts in a tick
  - This primarily improves the performance of `GetStateForCurrentFrame` which runs on EVERY tick
    - new - 3.04ms , old - 4.71ms

Before this Change
![JSON_AFTER](https://github.com/Regression-Games/RGUnityBots/assets/112959031/83f9536f-3fb6-450e-ae38-a8cba91b1039)

After this Change
![Dictionary_Prior_state_Perf](https://github.com/Regression-Games/RGUnityBots/assets/112959031/ca6376a6-9c09-4b32-8edd-3b34849cfbaf)


---

Find the pull request instructions [here](https://www.notion.so/regressiongg/Pull-Request-Process-0d57a7eb582a446983edfacafa406f1e?pvs=4)

Every reviewer and the owner of the PR should consider these points in their request (feel free to copy this checklist so you can fill it out yourself in the overall PR comment)

- [ ] The code is extensible and backward compatible
- [ ] New public interfaces are extensible and open to backward compatibility in the future
- [ ] If preparing to remove a field in the future (i.e. this PR removes an argument), the argument stays but is no longer functional, and attaches a deprecation warning. A linear task is also created to track this deletion task.
- [ ] Non-critical or potentially modifiable arguments are optional
- [ ] Breaking changes and the approach to handling them have been verified with the team (in the Linear task, design doc, or PR itself)
- [ ] The code is easy to read
- [ ] Unit tests are added for expected and edge cases
- [ ] Integration tests are added for expected and edge cases
- [ ] Functions and classes are documented
- [ ] Migrations for both up and down operations are completed
- [ ] A documentation PR is created and being reviewed for anything in this PR that requires knowledge to use
- [ ] Implications on other dependent code (i.e. sample games and sample bots) is considered, mentioned, and properly handled
- [ ] Style changes and other non-blocking changes are marked as non-blocking from reviewers